Add 3.17 builds, disable non-http feeds
Time to add 3.17 builders, with 3.17.1 pending.

rss, atom, and json feeds lock up twistd, presumably because there are
too many builders. This became apparent in the last few weeks and caused
the master to lock up frequently. Lockup seems to be triggered by web
accesses to those feeds. Either this a new problem due to added builders,
or someone started to actually use the feeds and is causing the lockup.
Either case, just disable the feeds to avoid potential denial of service
attacks and improve stability.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>