Drop PCI_MSI from x86 non-smp test configuration
This no longer works in kvm starting with a recent commit
in linux-next due to IO_APIC dependency changes. The failure
is blamed on qemu, which will take some time to get fixed,
if it will get fixed at all. Instead of carrying the failure,
drop PCI_MSI from the non-smp configuration to avoid a hang
while booting.

The SMP build is not affected since its IO_APIC configuration
in the kernel is different.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>