rootfs: Introduce common runtime infrastructure
If enabled, scripts in rootfs/scripts/runtime/ will be copied into
the target root file system. Scripts are expected to be common
among all architectures and can be managed outside the target code.

This will enable us to run selftests at a later time, and to simplify
script changes in general.

This currently only works for cpio archives.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>