alpha, mips: Drop CONFIG_SAMPLES
CONFIG_SAMPLES no longer builds (as of 4.1/kdbus) for alpha and mips
because of KDBUS expectations. Unfortunately, KDBUS sample builds can
not be disabled, so it is all or nothing. Also, there is a new expectation
that toolchains have to be built with kernel 3.17 or later, which we can not
guarantee. Therefore, drop CONFIG_SAMPLES from allmodconfig for alpha and
mips build to avoid unnecessary and useless build failures.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>