rootfs: ppc64: Add pseries to list of tests
Main difference is that it needs a different console. Also, the command
line _must_ include "-vga none", or the system will hang during boot.

pseries_defconfig works for the most part as-is. Only exception is that
DEVTMPFS has to be enabled explicitly for older kernel versions.

The pseries test also works for 3.2, so add skip entries for 3.2
to skip the other tests. The 3.2 test will be enabled in buildbot
with a separate commit.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>