mksdimage: Add support for new version of sfdisk command
sfdisk version 2.26 and later no longer supports the -D, -H, -S, and -C
options. Partition sizes now _must_ be specified in 512 byte sector sizes.
Make the necessary changes to support all versions of sfdisk.

The current code fixes the image size to 256M. This could be changed
by making the partition size of the linux partition dynamic,
similar to the original code, but for now it is good enough.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>