master: Rework parallel build support
Image builds keep a slave busy, so one build per slave makes sense.
For qemu test runs, however, it does not make sense to limit the number
of builds per slave to 1. This is ok for the image build, but the qemu
run itself results in substantial CPU idle time. Increate the number of
parallel qemu test runs on larger slaves to 3.

Change the qemu slave build directory to be per architecture instead of
per branch. With this change, the build system is fully utilized if a
single branch requires a rebuild (which is much more likely than
multi-branch rebuilds). Downside is that only one test per architecture
can run on a given slave, but that should not really matter much since
there are 20+ architectures to choose from.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>