master: Improve retry logic
We only want to retry if not all builds failed but at least
one build failed. We know this is the case if the build result
is a warning and at least one build failed (if all builds failed,
the build result would be a failure).

Downside is that the overall result after a retry still indicates
a warning. We may improve that later.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>