schedulers: TimedSingleBranchScheduler: Only upcall if no local timer
If a local timer is running, do not upcall but append the change to the
local change list (the timer should trigger shortly anyway). This ensures
that there are no scheduling conflicts against SingleBranchScheduler
at the end of the blockout time.

Note: There may be some overlap at the beginning of the blockout time if
a change is already pending. This would be difficult to resolve since the
time may move dynamically, so don't try it. We might expand the blockout
time by a configured treeStableTimer and have it start early, but that
doesn't really help either. Just accept that some builds may start even
if the blockout time has already started if they were queued earlier.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>