rootfs: s390: Boot bzImage instead of vmlinux
Quoting Vasily Gorbik:

"This change has been done on purpose. Uncompressed image is not going
to be bootable any more. In future the decompressor phase would get
more function (early memory detection as an example) and there is no
chance to duplicate that code in uncompressed image as well (to keep it
bootable on its own). The patch series commit messages contain more
technical details.

For qemu either bzImage or arch/s390/boot/compressed/vmlinux should be
used, which are bootable images."

Signed-off-by: Guenter Roeck <linux@roeck-us.net>