rootfs: alpha: Add FUSION, MEGASAS2 test builds
MEGASAS now uses megasas; MEGASAS2 uses megasas-gen2.

Enable nvme tests unconditionally; it appears those failed
due to interference with some other configuration option.

Add 53C810 and 53C895A to be built with -a. They fail, but maybe that
can be fixed at some point. It is unclear if the problem lies with
the driver or with qemu (though the driver does have a problem in
its error cleanup path; it disables interrupts where it should not).

Signed-off-by: Guenter Roeck <linux@roeck-us.net>