scripts: common: Fix broken fixup handling
Fixup handling needs access to .config to handle CONFIG_ATA vs. CONFIG_IDE
correctly, and must therefore run after the configuration file was created.
That step will only execute if the build wasn't cached, so add it to the
cache return path as needed and otherwise run it after configuration.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>