rootfs: sparc: Improve init.sh script for older CPUs
On older CPUs, the "Rebooting" message was interrupted by "random: fast init
done". As it turns out, that was systematic: The console output was used by
the random subsystem to provide randomness, and the amount of previous messages
(after adding debug logs) was just enough to _barely_ get enough randomness.
Displaying "Rebooting" was just enough to get over that limit, causing the
kernel to display the message.

Unfortunately, that disrupts the test script which expects to see uninterrupted
"Boot successful" and "Rebooting" messages. So display some nonsense until
"random: fast init done" is seen, the ncontinue with the actual reboot.
The init script in hda.sqf looks a bit odd since at the same time we have
to avoid buggy root fs code which otherwise tends to crash. Time to update
hda.sqf with a brand new root file system - if we ever manage to create one
that doesn't stall or crash on sparc v7 CPUs.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>