rootfs: arm: Fix sx1 boot from flash, and make it unconditional
The SX1 boot flash is really not 32MB but 8MB, and it doesn't start
at 0 but at 0x180000. Make up for that by specifying it as 32MB flash
with an offset of 24MB (26MB for the root file system). Use Squashfs
as root file system since only that is small enough to fit into the
6MB file system partition.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>