master: Skip hexagon build tests starting with v5.8
The minimum version of gcc is now v4.8, but there is no matching version
of gcc since hexagon only supports clang/llvm. On top of that, the current
version of llvm requires a library provided with llvm to build the kernel.

We don't (and won't) support llvm build tests, and we won't accept the
need for out-of-tree libraries to build the kernel. Drop hexagon build
tests instead.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>