rootfs: common: Truncate copy into flashes if incoming file is too large
We now have to define root file systems as exponent of 2
since qemu won't accept anything else for SD cards. That means
that the same root file system may now be too large to fit into
a flash if an offset (partition) into the flash is required.
Truncate the flash copy if that is the case since qemu won't
accept bad flash sizes either.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>