rootfs: scripts: Catch more crashes
In some cases, a crash results in an immediate qemu exit.
Check the kernel log for crashes afterwards to ensure that we catch
all crashes. That needs some extended tests for xtensa, which tends
to crash after reboot requests (most likey because the kernel just
jumps to the restart address but doesn't really reset the CPU).

Signed-off-by: Guenter Roeck <linux@roeck-us.net>