stable-build-arch: Disable building 6PACK on alpha
There is some disagreement about timeouts in the 6pack implementation
in the Linux kernel. The protocol specifies timeouts in increments of
10ms. The kernel implements timeouts relative to HZ. Timeout variables
and timeout packet data is kept in an u8. With a HZ value of 1024 or
larger, as used on alpha, this results in truncated timeout values and
with it, starting with v5.15, test build errors.

Attemts to fix the problem were unsuccessful as the maintainer states
that the driver should use HZ values, and also suggests to use unsigned
long instead of u8 for variables. This would silence the warning but
result in a runtime truncation. This means that 6pack fails to build
on alpha, and a solution is not in sight. To avoid useless build errors,
disable building 6PACK for alpha targets.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>