stable-build-arch: Use gcc 11.2 to build nds32 images
This requires a workaround: CONFIG_FUNCTION_TRACER and CONFIG_STACK_TRACER
need to be disabled. Not doing so results in assembler errors. For this toe
work, CONFIG_TRACE_IRQFLAGS needs to be disabled as well since it depends on
CONFIG_FUNCTION_TRACER, but that dependency is not spelled out in Kconfig.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>