rootfs: scripts: Set hung task timeout to 20 seconds
With the default 120 second timeout the qemu process is likely killed
before the hung task timeout is expired. Set the hung task timeout to
20 seconds to avoid that.

At the same time, disable the firmware download user helper.
We don't use it, and it has the side effect of blocking rtnl_lock
for up to 60 seconds if the firmware for some network devices
is not available (because those devices download the firmware from
the open function which is called under rtnl_lock).

Signed-off-by: Guenter Roeck <linux@roeck-us.net>