rootfs: arm-aspeed: Add boot tests from spi for bletchley and fuji-bmc
The trick was that both need to run with a memory size of 1G instead
of the default 2G to be able to instantiate the SPI interfaces.
Also, the default SPI flash on bletchley-bmc is not supported by
the upstream Linux kernel, so it was necessary to specify a supported
flash type.

Skip new fuji-bmc tests with v5.15.y.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>