rootfs: scripts: Fix usb-xhci
If usb-xhci is loaded in addition to a pre-existing USB bus, the device
expected to be attached to XHCI will be attached to the first usb bus
unless the bus is explicitly provided in the device command. This means
we have to explicitly explicitly attach devices to the xhci bus to ensure
that the device binds to it.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>