rootfs: scripts: Improve empty log line handling, crash cause detection
Detect empty log lines including tabs, and remove those from logs.

So far qemu failures were logged with "failed (network)" or similar,
which is misleading. Detect if qemu exited with an error. If so,
record as "failed (qemu)". While doing so, stop using "kill -0" to check
if a process is running; check for the existence of /proc/<pid>
instead.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>