stable-build-targets: Set FRAME_WARN limit to 4096 for x86_64
x86_64:allmodconfig as of v6.6-12893-ge392ea4d4d00 fails to build with:

Error log:
drivers/gpu/drm/amd/amdgpu/../display/dc/dml2/display_mode_core.c:
	In function 'dml_prefetch_check':
drivers/gpu/drm/amd/amdgpu/../display/dc/dml2/display_mode_core.c:6707:1:
	error: the frame size of 2056 bytes is larger than 2048 bytes

There was a discussion about this on mailing lists when this was observed
in linux-next, and there did not seem to be an agreement about a fix -
as is obvious since the problem is now seen in the mainline kernel.

If upstream doesn't care, we should not care either. It is more important
to get a passing build than to bother about frame size limits. Make the
frame size limit large enough to hopefully silence the warning/error for
us, and let others (if they are interested) deal with the problem, or not.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>