rootfs: arm64: Drop usb network tests for raspi3
The raspi3 USB interface fails to instantiate if the dwc2 interface
ends up in autosuspend/resume. Specifically, in _dwc2_hcd_resume(),
the hprt0 register has HPRT0_CONNSTS set which causes resume to fail.
Due to the resume failure, interrupts are (why ?) enabled while the
hardware is in inaccessible state (HCD_HW_ACCESSIBLE() returns false).
Since interrupts are ignored in this state, this results in interrupt
storms and subsequently missed interrupts, which later causes a hang
during reboot.

It is unknown if this is a qemu problem or a Linux kernel problem.
Either case, this results in random raspi3 test failures, to the point
where tests often have to be repeated manually which is just annoying and
does not add value.

Drop usb network tests for raspi3 to prevent the problem from being
reported.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>