rootfs: arm-aspeed: Stop testing second USB interface on AST2600
The second USB interface was supposed test UHCI, but on ast2600 the
UHCI interface is a companion of the EHCI interface and not supposed
to be visible as separate interface.

Test the second USB interface on ast2500-evb instead. AST2500 does not
support companion interfaces, so the second USB interface is a real UHCI
interface.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>