rootfs: arm-v7: Improve USB tests
The Allwinner H3 implementation does not select companion mode for its
OHCI interfaces. As result, the OHCI interfaces are created as separate
interfaces an can be tested explicitly. Test usb4 as OHCI interface.

For cubieboard, explicitly specify that the first (usb0) port is tested.

Also, use sqf for testing mtd32 flashes (ext2 is going to be too large
soon).

Signed-off-by: Guenter Roeck <linux@roeck-us.net>