shellcommands: Fix skip entries
If a skipped command was last, the entire build was counted as failure.
Fix it by setting the build result explicitly to either success if there are
passed builds with no failures or tracebacks, or warning if there are passed
builds with no failures but warnings.

For qemu commands, accept text after "skipped" to include cached results
and other details as skip reasons.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>