master: Reorder branch names
Buildbot does not always build branches if a child branch had
changes on top of not-yet pulled changes in the parent branch.

List branches in order of dependencies; maybe that helps to ensure
that pending-fixes actually builds, and that upstream builds even
if there was a subsequent change in testing.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>