rootfs: common: Disable SLUB_DEBUG and KFENCE for parisc 64-bit tests
SLUB_DEBUG enables STACKDEPOT which may be responsible for random
crashes when running the parisc64 (C3700) emulation. Similar, KFENCE
triggers a call to set_track_prepare which then may crash in
walk_stackframe(). It is unknown if there is something new in the
kernel, if this is an emulation problem, or if it is a real problem.
Either case, it is not worth trying to track it down. Disable the
problematic configurations instead.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>