rootfs: scripts: Use explicit MAC address for USB networking
If a local MAc address is provided (first MAC address byte has bit 1 set),
the Linux kernel driver creates network devices named usbX instead of ethX.
qemu used to set the first address byte to 0x40, so that was not a problem
However, a recent qemu change modified the first address byte, which is now
set to 0x52. This means that USB Ethernet interface now have an unexpected
name, which in turn causes test script failures.

Set an explicit non-local MAC address for USB network interfaces to
avoid the problem and to be qemu version independent.

Signed-off-by: Guenter Roeck <linux@roeck-us.net>